fix: Article Popular를 복구합니다
